refactor: move constants to core
