+++ packages/core/package.json
@@ -13,6 +13,16 @@
     "coverage": "vitest run --coverage",
     "lint": "eslint ."
   },
+  "exports": {
+    ".": {
+      "types": "./dist/index.d.ts",
+      "default": "./dist/index.js"
+    },
+    "./constants": {
+      "types": "./dist/constants.d.ts",
+      "default": "./dist/constants.js"
+    }
+  },
   "devDependencies": {
     "@junat/prettier": "workspace:*",
     "@junat/rollup": "workspace:^",
