+++ packages/core/src/constants.ts
@@ -1,13 +1,11 @@
-import { LOCALES } from './locales'
-
 export const SITE_NAME = 'Junat.live' as const
 export const SITE_DOMAIN = 'junat.live' as const
 export const SITE_URL = 'https://junat.live' as const
 
 export const DEFAULT_LOCALE = 'fi' as const
 
-// Use JavaScript to avoid transpiling TypeScript for use inside next.config.js
-export { LOCALES } from './locales'
+export const LOCALES = ['fi', 'en', 'sv'] as const
+
 export const INACTIVE_STATIONS = [
   'HSI',
   'HH',
@@ -26,6 +24,21 @@ export const FINTRAFFIC = {
   LICENSE_URL: 'https://creativecommons.org/licenses/by/4.0/deed.',
   LICENSE: 'CC 4.0 BY',
   LOCALE_PATHS: { fi: 'fi', sv: 'sv', en: 'en' },
+} as const
+
+export const ROUTES = {
+  fi: {
+    train: 'juna',
+    settings: 'asetukset',
+  },
+  en: {
+    train: 'train',
+    settings: 'settings',
+  },
+  sv: {
+    train: 'tog',
+    settings: 'installningar',
+  },
 } as const
 
 const constants = {
