+++ site/src/components/no_script/no_script.stories.tsx
@@ -2,7 +2,7 @@ import type { Meta, StoryFn } from '@storybook/react'
 import type { ComponentPropsWithoutRef } from 'react'
 import type { Locale } from '~/types/common'
 
-import { LOCALES } from '~/constants'
+import { LOCALES } from '@junat/core/constants'
 import { getSupportedLocale } from '~/i18n'
 import { translate } from '~/utils/translate'
 import { NoScript } from './'
