+++ site/src/features/pages/home/components/page.tsx
@@ -2,11 +2,12 @@ import type { GeolocationButtonProps } from '~/features/geolocation'
 import type { LocalizedStation } from '~/lib/digitraffic'
 import type { Locale } from '~/types/common'
 
-import React from 'react'
 import dynamic from 'next/dynamic'
 import Link from 'next/link'
 import { useRouter } from 'next/router'
+import React from 'react'
 
+import constants from '@junat/core/constants'
 import { Head } from '~/components/head'
 import { Header } from '~/components/header'
 import HeartFilled from '~/components/icons/heart_filled.svg'
@@ -14,7 +15,6 @@ import List from '~/components/icons/list.svg'
 import { Notification } from '~/components/notification'
 import { StationList } from '~/components/station_list'
 import { ToggleButton } from '~/components/toggle_button'
-import constants from '~/constants'
 import { getPrettifiedAccuracy } from '~/features/geolocation'
 import { SearchBar } from '~/features/search'
 import { useToast } from '~/features/toast'
