+++ site/src/lib/digitraffic/hooks/use_stations.ts
@@ -1,8 +1,8 @@
 import type { GetStationsOptions } from '@junat/digitraffic'
 import type { LocalizedStation } from '~/lib/digitraffic'
 
+import { INACTIVE_STATIONS } from '@junat/core/constants'
 import { useQuery } from '@tanstack/react-query'
-import { INACTIVE_STATIONS } from 'src/constants'
 
 import { fetchStations } from '@junat/digitraffic'
