+++ site/src/pages/_app.tsx
@@ -2,12 +2,12 @@ import type { AppProps as NextAppProps } from 'next/app'
 import type { ReactNode } from 'react'
 import type { LayoutProps } from '~/types/layout_props'
 
-import dynamic from 'next/dynamic'
-import { useRouter } from 'next/router'
 import { QueryClientProvider } from '@tanstack/react-query'
 import { ReactQueryDevtools } from '@tanstack/react-query-devtools'
+import dynamic from 'next/dynamic'
+import { useRouter } from 'next/router'
 
-import constants from '~/constants'
+import constants from '@junat/core/constants'
 import { useWakeLock } from '~/hooks/use_wake_lock'
 import { getSupportedLocale, LocaleProvider } from '~/i18n'
 import { queryClient } from '~/lib/react_query'
