+++ site/src/types/layout_props.ts
@@ -1,5 +1,5 @@
+import type constants from '@junat/core/constants'
 import type { ReactNode } from 'react'
-import type constants from 'src/constants'
 import type { Locale } from './common'
 
 export interface LayoutProps {
